requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.5
django-allauth==0.52.0
django-cors-headers==3.13.0
django-oauth-toolkit==2.2.0
django-otp==1.1.6
django-otp-twilio==1.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-social-oauth2==1.2.1
ecdsa==0.18.0
frozenlist==1.3.3
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
httplib2==0.21.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
multidict==6.0.4
oauthlib==3.2.2
Pillow==9.4.0
pip-chill==1.0.1
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
twilio==8.0.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.15.0
yarl==1.8.2
